move networkflow codes into a new file
